requirements.txt
python-telegram-bot
google-genai
edge-tts
supabase
python-dotenv
imageio-ffmpeg
flask